Design -> fixup auth padding-top, modularize Tasks backend, fixup seeds, color headings
