+++ app/Http/Controllers/Tasks/TaskController.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Tasks;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use App\Task;
+use App\Models\Tasks\Task;
 
 class TaskController extends Controller
 {
