+++ app/Models/Tasks/Task.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models\Tasks;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -8,13 +8,15 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 class Task extends Model
 {
     use SoftDeletes;
-    
+
+    protected $dates = ['deleted_at'];
+
     protected $fillable = [
         'title', 'done', 'order', 'user_id',
     ];
 
     public function user()
     {
-      return $this->belongsTo(User::class);
+      return $this->belongsTo(\App\User::class);
     }
 }
