+++ resources/assets/sass/_variables.scss
@@ -15,6 +15,7 @@ $hovered: #f9fafb;
 
 // Typography
 $text-color: darken($brand-gray, 20%);
+$headings-color: darken($brand-gray, 10%);
 
 // mdi icons
 .mdi-10x, .mdi-size-10x {
