+++ resources/assets/sass/modules/_auth.scss
@@ -1,6 +1,7 @@
 .auth{
     padding: 0;
     margin: 0;
+    margin-top: -65px;
     .left-pane{
         position: relative;
         float: left;
